move handler app tests to manifest app
