--- server/handler/models/__init__.py
@@ -1,3 +1,0 @@
-"""
-Model definitions for the trak domain
-"""
