+++ server/manifest/tests/test_handler_models.py
@@ -1,4 +1,5 @@
 import pytest
+
 from manifest.models import Handler
