+++ server/manifest/tests/test_handler_serializer.py
@@ -1,4 +1,5 @@
 import pytest
+
 from manifest.models import Handler, PaperSignature
 from manifest.serializers import HandlerSerializer
