+++ server/manifest/tests/test_handler_services.py
@@ -1,5 +1,4 @@
 import pytest
-
 from core.services import RcraClient
 from rcrasite.models import RcraSite
 from rcrasite.services import RcraSiteService
